requirements.txt
anthropic
alfred-workflow